use-case-exp/iweblens/client/requirements.txt
locust==2.16.*+
PyYAML